[TE-5476] Extract FakeGitRunner to shared file for cross-package reuse
Move FakeGitRunner from git_test.go to internal/git/fake_runner.go
(non-test file under internal/) so it can be imported by tests in
other packages. Remove the duplicate fakeGitRunner from the command
tests and use git.FakeGitRunner directly.
